+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Secrets (`rules configs` and databases `options.configuration`) can no longer be exported
+
 ## [3.1.2] - 2019-03-22
 ### Added
 - Consistent property sorting for yaml dump #108 #61 #82